move feature flags into their own file
