+++ src/lib/store/ConfigModel.ts
@@ -2,82 +2,10 @@ import { action, Action, computed, Computed, thunk, Thunk, thunkOn, ThunkOn } fr
 import moment from "moment"
 import Config from "react-native-config"
 import echoLaunchJSON from "../../../Artsy/App/EchoNew.json"
+import { FeatureName, features } from "./features"
+
 import { GlobalStoreModel } from "./GlobalStoreModel"
 
-interface FeatureDescriptor {
-  /**
-   * Set readyForRelease to `true` when the feature is ready to be exposed outside of dev mode.
-   * Echo flags and gravity lab features still take precedence after this is set to `true`.
-   * If this is set to `false`, the feature will never be shown except if overridden in the admin menu.
-   */
-  readonly readyForRelease: boolean
-  /**
-   * Provide an echo feature flag key to allow this feature to be toggled globally via echo.
-   */
-  readonly echoFlagKey?: string
-  /**
-   * Provide a description to show the feature flag in the admin menu
-   */
-  readonly description?: string
-}
-
-// Helper function to get good typings and intellisense
-function defineFeatures<T extends string>(featureMap: { readonly [featureName in T]: FeatureDescriptor }) {
-  return featureMap
-}
-export const features = defineFeatures({
-  AROptionsBidManagement: {
-    readyForRelease: true,
-    echoFlagKey: "AROptionsBidManagement",
-  },
-  AROptionsArtistSeries: {
-    readyForRelease: true,
-    echoFlagKey: "AROptionsArtistSeries",
-  },
-  AROptionsNewFirstInquiry: {
-    readyForRelease: true,
-    echoFlagKey: "AROptionsNewFirstInquiry",
-  },
-  AROptionsNewInsightsPage: {
-    readyForRelease: false,
-    echoFlagKey: "AROptionsNewInsightsPage",
-  },
-  AROptionsInquiryCheckout: {
-    readyForRelease: false,
-    description: "Enable inquiry checkout",
-  },
-  AROptionsPriceTransparency: {
-    readyForRelease: true,
-    echoFlagKey: "AROptionsPriceTransparency",
-    description: "Price Transparency",
-  },
-  ARDisableReactNativeBidFlow: {
-    readyForRelease: true,
-    echoFlagKey: "ARDisableReactNativeBidFlow",
-  },
-  AREnableNewPartnerView: {
-    readyForRelease: true,
-    echoFlagKey: "AREnableNewPartnerView",
-  },
-  AROptionsUseReactNativeWebView: {
-    readyForRelease: false,
-    description: "Use react-native web views",
-  },
-  AROptionsLotConditionReport: {
-    readyForRelease: true,
-    echoFlagKey: "AROptionsLotConditionReport",
-  },
-  AROptionsNewSalePage: {
-    readyForRelease: true,
-    echoFlagKey: "AROptionsNewSalePage",
-  },
-  AREnableViewingRooms: {
-    readyForRelease: true,
-    echoFlagKey: "AREnableViewingRooms",
-  },
-})
-
-export type FeatureName = keyof typeof features
 export type FeatureMap = { [k in FeatureName]: boolean }
 
 type EchoJSON = typeof echoLaunchJSON
