+++ src/lib/store/GlobalStore.tsx
@@ -5,7 +5,8 @@ import React from "react"
 import { Platform } from "react-native"
 import Config from "react-native-config"
 import { Action, Middleware } from "redux"
-import { FeatureMap, FeatureName, features } from "./ConfigModel"
+import { FeatureMap } from "./ConfigModel"
+import { FeatureName, features } from "./features"
 import { GlobalStoreModel, GlobalStoreState } from "./GlobalStoreModel"
 import { assignDeep, persistenceMiddleware, unpersist } from "./persistence"
